requirements.txt
asgiref==3.7.2
Django==4.2.2
djangorestframework==3.14.0
greenlet==2.0.2
mysqlclient==2.2.0
numpy==1.25.0
pandas==2.0.2
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.17
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
